extracted useCustomDialog to the hooks folder since it can be used in VideoSuite.tsx
