+++ src/hooks/useCustomDialog.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react"
 
 const useCustomDialog = ({ openFromProps, freeze }: Props) => {
-	const [open, setOpen] = useState(openFromProps ? openFromProps : false)
+	const [open, setOpen] = useState(false)
 
 	useEffect(() => {
 		if (openFromProps !== undefined) {
